+++ src/server/live-error.js
@@ -1,0 +1,15 @@
+// @ts-check
+
+/**
+ * Typed error that propagates code to the client.
+ */
+export class LiveError extends Error {
+	/**
+	 * @param {string} code
+	 * @param {string} [message]
+	 */
+	constructor(code, message) {
+		super(message || code);
+		this.code = code;
+	}
+}